Comparing Cust K_Means and SK learn
Here I am comparing the results of my analysis on the titanic data set, and SK learns analysis, which may be more thoughtful. However, I ran into a minor hitch when converting non-numerical data, the convert_objects operator has been depreciated by pandas, and this makes it difficult. It will most likely work on an older version, or with Py3.7